backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.11.4
psycopg2-binary